On Go 1.21 or later, DetachContextSegment is just an alias of context.WithoutCancel.
