+++ xray/segment.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"net/http"
 	"os"
-	"reflect"
 	"strings"
 	"sync"
 	"time"
@@ -762,42 +761,4 @@ func (seg *Segment) AddAnnotationFloat64(key string, value float64) {
 // AddAnnotationFloat64 adds a 64 bit integer type annotation.
 func AddAnnotationFloat64(ctx context.Context, key string, value float64) {
 	ContextSegment(ctx).addAnnotation(key, value)
-}
-
-// DetachContextSegment returns a new context with the existing segment.
-// All values associated with ctx are also associated with the new context.
-// This is useful for creating background tasks which won't be cancelled
-// when a request completes.
-func DetachContextSegment(ctx context.Context) context.Context {
-	return &detachedContext{Context: ctx}
-}
-
-type detachedContext struct {
-	context.Context
-}
-
-func (*detachedContext) Deadline() (deadline time.Time, ok bool) {
-	return
-}
-
-func (*detachedContext) Done() <-chan struct{} {
-	return nil
-}
-
-func (*detachedContext) Err() error {
-	return nil
-}
-
-func (ctx *detachedContext) String() string {
-	return contextName(ctx.Context) + ".Detached"
-}
-
-func contextName(c context.Context) string {
-	type stringer interface {
-		String() string
-	}
-	if s, ok := c.(stringer); ok {
-		return s.String()
-	}
-	return reflect.TypeOf(c).String()
 }
